switched from int to float for min/max dmg fixed warden right click animation fixed earth spike speed diagonally (looks weird tho) fixed ground slam being stuck fixed sg totem when dying fixed intervene not breaking when defender dies added weapon skill boosts (weapon description updates when game starts but numbers for skills dont)

+++ src/com/ebicep/warlords/classes/abilties/Breath.java
@@ -16,7 +16,7 @@ import java.util.List;
 
 public class Breath extends AbstractAbility {
 
-    public Breath(String name, int minDamageHeal, int maxDamageHeal, int cooldown, int energyCost, int critChance, int critMultiplier, String description) {
+    public Breath(String name, float minDamageHeal, float maxDamageHeal, int cooldown, int energyCost, int critChance, int critMultiplier, String description) {
         super(name, minDamageHeal, maxDamageHeal, cooldown, energyCost, critChance, critMultiplier, description);
     }
