+++ src/com/ebicep/warlords/classes/abilties/Consecrate.java
@@ -16,15 +16,15 @@ import org.bukkit.scheduler.BukkitTask;
 import java.util.List;
 
 public class Consecrate extends AbstractAbility {
-    public Consecrate(int minDamageHeal, int maxDamageHeal, int energyCost, int critChance, int critMultiplier) {
+    public Consecrate(float minDamageHeal, float maxDamageHeal, int energyCost, int critChance, int critMultiplier) {
         super("Consecrate", minDamageHeal, maxDamageHeal, 8, energyCost, critChance, critMultiplier,
                 "§7Consecrate the ground below your\n" +
-                "§7feet, declaring it sacred. Enemies\n" +
-                "§7standing on it will take §c%dynamic.value% §7-\n" +
-                "§c%dynamic.value% §7damage per second and\n" +
-                "§7take §c%dynamic.value% §7increased damage from\n" +
-                "§7your paladin strikes. Lasts §65\n" +
-                "§7seconds.");
+                        "§7feet, declaring it sacred. Enemies\n" +
+                        "§7standing on it will take §c%dynamic.value% §7-\n" +
+                        "§c%dynamic.value% §7damage per second and\n" +
+                        "§7take §c%dynamic.value% §7increased damage from\n" +
+                        "§7your paladin strikes. Lasts §65\n" +
+                        "§7seconds.");
     }
 
     @Override
