+++ src/com/ebicep/warlords/classes/abilties/Strike.java
@@ -22,7 +22,7 @@ import java.util.List;
 
 public class Strike extends AbstractAbility {
 
-    public Strike(String name, int minDamageHeal, int maxDamageHeal, int cooldown, int energyCost, int critChance, int critMultiplier, String description) {
+    public Strike(String name, float minDamageHeal, float maxDamageHeal, int cooldown, int energyCost, int critChance, int critMultiplier, String description) {
         super(name, minDamageHeal, maxDamageHeal, cooldown, energyCost, critChance, critMultiplier, description);
     }
